+++ CHANGES.md
@@ -2,6 +2,19 @@
 
 Changes in descending order.
 
+### 10.06.2018 (v1.0.24)
+
+<table>
+  <tr><td>exit</td><td>flush queue for lokales when debugging rather than waiting to empty.</td></tr>
+  <tr><td>core.dispatch</td><td>if terminate enabled and is aborting halt further dispatches.</td></tr>
+</table>
+
+### 09.22.2018 (v1.0.23)
+
+<table>
+  <tr><td>error</td><td>favor full stacktrace rather than pruned.</td></tr>
+</table>
+
 ### 09.20.2018 (v1.0.19-v1.0.22)
 
 <table>